trying to mess with the directoy structure, not rerally sure what im doing
