+++ src/scripts/directory.ps1
@@ -1,0 +1,1 @@
+Set-Location -Path "C:\Users\godih\OneDrive\Documents\NOC\fall-2024\cpp\final-project\cpp-http-server\"
